[refactor] feat:main 모듈 추가함으로써 app모듈에서 ui관련 코드 제거
