+++ core/network/booksearch/build.gradle.kts
@@ -23,5 +23,9 @@ android {
 
 dependencies {
 	implementation(project(":core:data:booksearch:external"))
+	implementation(project(":core:network:network"))
 	implementation(libs.bundles.ktor)
 }
+//드로이드 나이츠처럼 app 모듈이 모든 모듈을 참조해야하는가? (나는 의존성 주입을 위해 그런줄 알았는데 왜 드로이드 나이츠는 저런 구조로도 가능한건가)
+//datastore:searchhistory는 datastore:datastore 모듈을 참조하는데 네트워크 모듈을 왜 다른가?
+//datastore:datastore, network:network 모듈 처럼 해당 기능을 위한 모듈이 필요한게 맞는가
