+++ feature/main/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+	id("convention.android.feature")
+}
+
+android {
+	namespace = "com.kova700.amazonbookstorepractice.feature.main"
+}
+
+dependencies {
+	implementation(project(":feature:search"))
+}
