+++ feature/main/src/main/kotlin/com/kova700/feature/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.kova700.amazonbookstorepractice
+package com.kova700.feature.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
